requirements.txt
flask==3.0.3
requests==2.32.3
pymongo==4.8.0
python-consul==1.1.0
Werkzeug==3.0.2
itsdangerous==2.2.0
click==8.1.7